AdminServiceをインターフェースと実装クラス(AdminServiceImpl)に分離
