+++ practice/src/main/java/com/example/demo/service/AdminService.java
@@ -2,50 +2,13 @@ package com.example.demo.service;
 
 import java.util.Optional;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-
 import com.example.demo.entity.Admin;
 import com.example.demo.form.AdminForm;
 import com.example.demo.form.SigninForm;
-import com.example.demo.repository.AdminRepository;
 
-@Service
-public class AdminService {
-	@Autowired
-	private AdminRepository adminRepository;
+public interface AdminService {
+	void signup(AdminForm  form);
 	
-	@Autowired
-	private PasswordEncoder passwordEncoder;
-	
-	public void signup(AdminForm form) {
-		
-		if (adminRepository.existsByEmail(form.getEmail())) {
-            // すでに存在する場合は、例外を投げて処理を中断する
-            throw new RuntimeException("Duplicate entry for email");
-        }
-		
-		Admin admin = new Admin();
-		admin.setLastName(form.getLastName());
-		admin.setFirstName(form.getFirstName());
-		admin.setEmail(form.getEmail());
-		admin.setPassword(passwordEncoder.encode(form.getPassword()));  
-		
-		adminRepository.save(admin);
-	}
-	
-	public Optional<Admin> signin(SigninForm form){
-		Admin admin = adminRepository.findByEmail(form.getEmail());
-		if (admin == null) {
-			return Optional.empty();
-		}
-		
-		if (passwordEncoder.matches(form.getPassword(),admin.getPassword())) {
-			return Optional.of(admin);
-		}
-		
-		return Optional.empty();
-	}
-
+	Optional<Admin> signin(SigninForm form);
 }
+
