+++ practice/src/main/resources/templates/contactList.html
@@ -27,7 +27,7 @@
 			</td>
 		</tr>
 	</table>
-	<form th:action="@{/admin/signin}" method="get">
+	<form th:action="@{/admin/logout}" method="post">
 			<button type="submit">ログアウト</button>
 		 </form>		
 </body>
