+++ practice/src/main/resources/templates/detail.html
@@ -38,7 +38,7 @@
 			</tr>
 		</table>
 		<a th:href="@{/admin/contacts/{id}/edit(id=${contact.id})}">編集画面へ</a>
-		<form th:action="@{/admin/signin}" method="get">
+		<form th:action="@{/admin/logout}" method="post">
 				<button type="submit">ログアウト</button>
 			 </form>
