+++ practice/src/main/resources/templates/edit.html
@@ -84,7 +84,7 @@
 	 <form th:action="@{/admin/contacts/{id}(id=${contactId})}" method="post">	
 	 		<button type="submit">削除</button>
 	 </form>
-	 <form th:action="@{/admin/signin}" method="get">
+	 <form th:action="@{/admin/logout}" method="post">
 		<button type="submit">ログアウト</button>
 	 </form>
