+++ practice/src/main/resources/templates/welcome.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="UTF-8">
+<title>管理者ページ</title>
+</head>
+<body>
+<h1>管理者ページ</h1>
+<p th:text="'ログイン: ' + ${admin.email}"></p>
+
+<form th:action="@{/admin/logout}" method="post">
+        <button type="submit">ログアウト</button>
+    </form>
+</body>
+</html>
